Add pysmt solvers installer wrapper to maxgap setup
A list of solvers (z3 for now only) is installed during
`penaltymodel_maxgap` setup, converting what used to be a manual
process for users to a one-line pip install step.

The list of solvers can still be checked and/or updated with
`pysmt-install` command.

Note: since most of the solvers (including z3) are built from C on
install, we can't have `penaltymodel_maxgap` available as bdist_wheel.
Source distribution is the only safe option. Also, no more `zip_safe`.